fixup! Replaced math.h double functions by their float equivalent

+++ kernels/volk/volk_32f_acos_32f.h
@@ -135,7 +135,7 @@ volk_32f_acos_32f_a_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = acos(*aPtr++);
+    *bPtr++ = acosf(*aPtr++);
   }
 }
 
@@ -206,7 +206,7 @@ volk_32f_acos_32f_u_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = acos(*aPtr++);
+    *bPtr++ = acosf(*aPtr++);
   }
 }
 
@@ -222,7 +222,7 @@ volk_32f_acos_32f_generic(float* bVector, const float* aVector, unsigned int num
   unsigned int number = 0;
 
   for(number = 0; number < num_points; number++){
-    *bPtr++ = acos(*aPtr++);
+    *bPtr++ = acosf(*aPtr++);
   }
 
 }
